Separate memorable_date into separate file
